added spacing to QuizResults elements, renamed PopupMenu
