+++ src/components/Quiz/Quiz.css
@@ -43,3 +43,7 @@
     justify-content: center;
     padding: 2px;
 }
+
+.quiz-results-div-item {
+    margin: 5px;
+}
