+++ src/components/Quiz/QuizResults.jsx
@@ -1,9 +1,10 @@
 import React, {useState} from "react";
 import "../Base.css";
 import "./Quiz.css";
-import PopupMenu from "../PopupMenu";
-import PopupMenuItem from "../PopupMenuItem";
+import Popup from "../Popup";
+import PopupItem from "../PopupItem";
 import { useNavigate } from "react-router-dom";
+
 
 
 
@@ -44,17 +45,17 @@ const QuizResults = (props) => {
     }
 
     return (     
-        <div className="quiz-results-div">
-            <p> {props.data.points}/{props.questions.length}pts | {props.data.date} </p>
-            <button className="btn-orange" onClick={handlePopupButtonClick}>Try Again?</button>
-            <button className="btn-orange-outline" onClick={() => navigate("/history")}>See Attempts History</button>
+        <>
+            <p className="quiz-results-div-item"> {props.data.points}/{props.questions.length}pts | {props.data.date} </p>
+            <button className="quiz-results-div-item btn-orange" onClick={handlePopupButtonClick}>Try Again?</button>
+            <button className="quiz-results-div-item btn-orange-outline" onClick={() => navigate("/history")}>See Attempts History</button>
             {showPopup && 
-            <PopupMenu setShowPopup={setShowPopup}>
+            <Popup setShowPopup={setShowPopup}>
                 <p>Do you want to save this attempt?</p>
-                <PopupMenuItem onClick={handleSaveHistory}>Yes</PopupMenuItem>
-                <PopupMenuItem onClick={() => navigate("/")}>No</PopupMenuItem>
-            </PopupMenu>}
-        </div>
+                <PopupItem onClick={handleSaveHistory}>Yes</PopupItem>
+                <PopupItem onClick={() => navigate("/")}>No</PopupItem>
+            </Popup>}
+        </>
     );
 }
